// Centralizador-Log-main/frontend-angular/src/app/core/models/api.models.ts
export interface AuthUserProfile {
    username: string;
    display_name: string;
    email: string;
}

export interface AuthResponse {
    success: boolean;
    message: string;
    user?: AuthUserProfile;
    access_token?: string;
    token_type?: string;
    expires_in?: number;
    expires_at?: string;
}